Tactics died before it existed on any club in the 45bn band
The world rebuild sets game_pid = id for every curated and generated record — 306,185 of
401,149 players on the live save, every one past Int32 — and PlayerRow.GamePid was an int.
Dapper mapping an int64 column into an int property does not truncate; it throws. So
Repository.SquadPlayers threw OverflowException for Liverpool and Arsenal, the two career
clubs in that band, and every screen built on Session.Squad() — Tactics first — was dead
before it was constructed. The crash log on the manager's machine shows it four times in two
seconds: 'Error parsing column 1 (GamePid=45000000001 - Int64)'.

It also meant the opening-XI seed could never have run on that save: SuggestXi reads the
squad through the same call, threw, and the constructor's catch logged it and moved on. Every
verification pass measured the live database and reasoned about id columns; none of them
exercised Dapper's mapping of game_pid, and the smoke career (Bournemouth, 20M-band ids with
small game ids) could not show it. It can now: the smoke gives its squad 45bn game ids and
proves the squad reads and the assistant can still order it. 56/56.

GamePid is a long. The match-event tuple's PlayerId is a long for the same reason. The two
write sites that cast to int no longer do; nothing on the C# side reads game_pid into an int.

+++ src/ML.App/SessionWorld.cs
@@ -792,7 +792,7 @@ public sealed partial class Session
                 {
                     // Id is the handle and is long; game_pid is an int column, and the academy
                     // band is bounded below 40m precisely so this cast is always exact.
-                    Id = id, GamePid = (int)id, IsCustom = true, Name = name, Position = position,
+                    Id = id, GamePid = id, IsCustom = true, Name = name, Position = position,
                     Age = age, Nationality = club.ShortName, OverallRating = rating,
                 });
                 // Real potential: most are journeymen; ~1 in 9 is a genuine prospect.
@@ -867,7 +867,7 @@ public sealed partial class Session
             var id = r.GetInt64(0);
             rows.Add(new PlayerRow
             {
-                Id = id, GamePid = (int)id, IsCustom = true, Name = r.GetString(1),
+                Id = id, GamePid = id, IsCustom = true, Name = r.GetString(1),
                 Position = r.IsDBNull(2) ? "CMF" : r.GetString(2),
                 Age = r.IsDBNull(3) ? null : r.GetInt32(3),
                 OverallRating = r.IsDBNull(4) ? null : r.GetInt32(4),
